uniform debug :(3

+++ models/modded_basic_nflow.py
@@ -585,8 +585,8 @@ def create_mixture_flow_model(input_dim, context_dim, base_kwargs, transform_typ
     """
 
     distribution = distributions.uniform.BoxUniform(
-        low=torch.full((input_dim,), -3, dtype=float),
-        high=torch.full((input_dim,), 3, dtype=float),
+        low=torch.full((input_dim,), -3, dtype=torch.float32),
+        high=torch.full((input_dim,), 3, dtype=torch.float32),
     )
     transform = []
     for _ in range(base_kwargs["num_steps_maf"]):
